\name{oglmx-package}
\alias{oglmx-package}
%\alias{oglmx}
\docType{package}
\title{
\packageTitle{oglmx} Package for estimation of ordered generalized linear models.
}
\description{
\packageDescription{oglmx}
}
\details{
%The DESCRIPTION file:
\packageDESCRIPTION{oglmx}
\packageIndices{oglmx}
%~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{oglmx}

Maintainer: \packageMaintainer{oglmx}
}
%\references{
%~~ Literature or other references for background information ~~
%}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
%\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
